requirements.txt
fastapi
uvicorn
dnspython
pymongo
motor
aiohttp
pynacl
discord-webhook
discord.py
cool-utils